requirements.txt
polars
duckdb
fredapi
pandas
statsmodels
matplotlib
python-dotenv
dbt-duckdb
pyarrow
mlflow
google-cloud-storage
gcsfs
plotly
streamlit